add_caption_cwb

Add a social media caption to a ggplot or gt table.

Usage

add_caption_cwb(
  type = c("plot", "table"),
  include_data_source = FALSE,
  data_source = NULL,
  github_username = "colebaril",
  bluesky_username = "@colebaril.ca"
)

Arguments

type

Either "plot" or "table"

github_username

Your GitHub username (default: "colebaril")

x

A ggplot object or gt table

Value

A ggplot or gt object with a caption applied